% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utilities.R
\name{add_seurat_embedding}
\alias{add_seurat_embedding}
\title{Easily add new embeddings to Seurat Object}
\usage{
add_seurat_embedding(object, embedding, reduction_name, type = "umap")
}
\arguments{
\item{embedding}{a matrix containing the embedding.}

\item{reduction_name}{name to save the reduction as}

\item{type}{embedding type (e.g. umap (default), 'harmony','hdbscan','pca')}
}
\description{
Easily add new embeddings to Seurat Object
}
