% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_utilities.R
\name{cellphonedb_summary}
\alias{cellphonedb_summary}
\title{CellPhoneDB Summary File}
\usage{
cellphonedb_summary(path, pvalue = "all")
}
\arguments{
\item{path}{the directory containing the CellPhoneDB Output}

\item{pvalue}{setting this will return results less than it}
}
\description{
CellPhoneDB Summary File
}
