% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{clustUMAP}
\alias{clustUMAP}
\title{UWOT-UAMP: Clustering Specific UMAP}
\usage{
clustUMAP(
  object,
  reduction = "harmony",
  spread = 1.1,
  n_components = NULL,
  min_dist = 0,
  metric = "cosine",
  n_neighbors = 50,
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  repulsion_strength = 1,
  negative_sample_rate = 5,
  n_threads = parallel::detectCores() - 1,
  reduction_name = "umap",
  return_seurat = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{whether to print function messages}
}
\description{
UWOT-UAMP: Clustering Specific UMAP
}
