% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_helpers.R
\name{create_seurat}
\alias{create_seurat}
\title{Simple method for creating Seurat Objects}
\usage{
create_seurat(filepath, sample = NULL)
}
\arguments{
\item{sample}{sample name to use}
}
\value{

}
\description{
Simple method for creating Seurat Objects
}
