% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_utilities.R
\name{db_mtx_remove_empty}
\alias{db_mtx_remove_empty}
\title{Remove empty columns from sparse matrix}
\usage{
db_mtx_remove_empty(M, ...)
}
\arguments{
\item{...}{}
}
\description{
Remove empty columns from sparse matrix
}
