% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_qc.R
\name{db_qc}
\alias{db_qc}
\title{Add QC metrics to Seurat Object}
\usage{
db_qc(
  object,
  species = "Mus musculus",
  subset = FALSE,
  mito = 10,
  ribo = 2.5,
  Complexity = 0.8,
  min.features = 500,
  cell_cycle = TRUE
)
}
\arguments{
\item{min.features}{}
}
\description{
Add QC metrics to Seurat Object
}
