% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read10x.R
\name{db_read10x}
\alias{db_read10x}
\title{Fast Import of 10X Data}
\usage{
db_read10x(
  path,
  return = "Seurat",
  gene_col = 2,
  barcode_col = 1,
  seurat_project = basename(path),
  QC = TRUE,
  trim_fat = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{return}{Set the format for the imported data. By default, 10X data is imported as a sparse matrix (\code{return='sparse'}), yet it can also return it as either a \code{SingleCellExperiment} (\code{return='sce'}) or  \code{SeuratObject} (\verb{return='Seurat}).}

\item{gene_col}{The column position to}

\item{trim_fat}{}
}
\description{
Fast Import of 10X Data
}
