% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read10x.R
\name{db_read_alevin}
\alias{db_read_alevin}
\title{Read Count data from Salmon Alevin}
\usage{
db_read_alevin(path = ".", return = "Seurat")
}
\arguments{
\item{return}{}
}
\description{
Read Count data from Salmon Alevin
}
