% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read10x.R
\name{db_read_mtx}
\alias{db_read_mtx}
\title{Import \code{.mtx} files}
\usage{
db_read_mtx(mtx, features, cells, return = "matrix", transpose_mtx = FALSE)
}
\arguments{
\item{transpose_mtx}{}
}
\description{
Import \code{.mtx} files
}
