% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_10x.R
\name{db_write10X}
\alias{db_write10X}
\title{Write 10X Files from Sparse Matrix}
\usage{
db_write10X(matrix, path = "output", gene_id = NULL, gene_symbol = NULL)
}
\arguments{
\item{gene_symbol}{}
}
\description{
Write 10X Files from Sparse Matrix
}
