% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read10x_experimental.R
\docType{class}
\name{dbsc-class}
\alias{dbsc-class}
\alias{dbsc}
\title{\code{dbsc} data class}
\description{
\code{dbsc} data class
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table.}

\item{\code{genes}}{character.}

\item{\code{barcodes}}{character.}
}}

