% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read10x.R
\docType{class}
\name{dbsc2-class}
\alias{dbsc2-class}
\alias{dbsc2}
\title{\code{dbsc2} data class}
\description{
\code{dbsc2} data class
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table.}

\item{\code{genes}}{data.table.}

\item{\code{barcodes}}{data.table.}
}}

