% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_10x.R
\name{flatMatrix_2_10x}
\alias{flatMatrix_2_10x}
\title{Convert a flat text file into a 10X-formated folder}
\usage{
flatMatrix_2_10x(file, new_folder = NULL, extract = NULL)
}
\arguments{
\item{new_folder}{}
}
\description{
Convert a flat text file into a 10X-formated folder
}
