% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{hcl_umap}
\alias{hcl_umap}
\title{UMAP Palette using HCL presets}
\usage{
hcl_umap(
  object,
  group_col,
  hcl_pal = "Dark 3",
  alpha = 0.7,
  jitter = TRUE,
  comp = 3
)
}
\arguments{
\item{comp}{integer setting the color complementary to be used}
}
\description{
UMAP Palette using HCL presets
}
