% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan.R
\name{HDBSCAN.default}
\alias{HDBSCAN.default}
\title{Hierarchical Density-Based Spatial Clustering of Applications with Noise}
\usage{
\method{HDBSCAN}{default}(
  x,
  algorithm = "best",
  alpha = 1,
  approx_min_span_tree = TRUE,
  gen_min_span_tree = FALSE,
  leaf_size = 30,
  metric = "euclidean",
  prediction_data = TRUE,
  min_cluster_size = 1000L,
  min_samples = 1L,
  cluster_selection_epsilon = 0.1,
  cluster_selection_method = "leaf",
  nThreads = parallel::detectCores() - 1,
  return_full = FALSE
)
}
\arguments{
\item{nThreads}{}
}
\description{
Hierarchical Density-Based Spatial Clustering of Applications with Noise
}
