% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan.R
\name{HDBSCAN.Seurat}
\alias{HDBSCAN.Seurat}
\title{Hierarchical Density-Based Spatial Clustering of Applications with Noise}
\usage{
HDBSCAN.Seurat(
  object,
  reduction = "umap",
  dims = NULL,
  algorithm = "best",
  alpha = 1,
  prediction_data = TRUE,
  approx_min_span_tree = TRUE,
  gen_min_span_tree = FALSE,
  leaf_size = 40,
  metric = "euclidean",
  min_cluster_size = 50,
  min_samples = 1,
  cluster_selection_epsilon = 0.5,
  cluster_selection_method = "leaf",
  nThreads = parallel::detectCores(),
  return_seurat = TRUE
)
}
\arguments{
\item{prediction_data}{not sure what this is for. Will update later.}

\item{return_seurat}{logical to return the result within the orignal object or as the raw HDBSCAN result}
}
\description{
Hierarchical Density-Based Spatial Clustering of Applications with Noise
}
