% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdb.R
\name{import_cpdb_results}
\alias{import_cpdb_results}
\title{Simplified importing CellphoneDB Files}
\usage{
import_cpdb_results(path)
}
\arguments{
\item{path}{}
}
\description{
Simplified importing CellphoneDB Files
}
