% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_utilities.R
\name{message_section}
\alias{message_section}
\title{Internal Function Messages}
\usage{
message_section(text, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Internal Function Messages
}
