% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle_processing.R
\name{monocle_processing}
\alias{monocle_processing}
\title{Monocle3 workflow for the lazy man}
\usage{
monocle_processing(object, dims = 100, verbose = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{dims for preprocess}

\item{verbose}{Default = TRUE}
}
\description{
Monocle3 workflow for the lazy man
}
