% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewMeta.R
\name{NewMeta}
\alias{NewMeta}
\title{Easy Add Meta data to Seurat Object}
\usage{
NewMeta(object, meta, col.name)
}
\arguments{
\item{col.name}{name of the column for the new meta data}
}
\value{

}
\description{
Easy Add Meta data to Seurat Object
}
