% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_10x.R
\name{organize_10x}
\alias{organize_10x}
\title{Reorganize another person's mess into a usable 10X dataset}
\usage{
organize_10x(x)
}
\arguments{
\item{x}{path containing the unorganized disaster}
}
\value{

}
\description{
Reorganize another person's mess into a usable 10X dataset
}
