% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{PaCMAP.default}
\alias{PaCMAP.default}
\title{Pairwise Controlled Manifold Approximation (R wrapper)}
\usage{
\method{PaCMAP}{default}(
  x,
  n_components = 2,
  n_neighbors = 10,
  MN_ratio = 0.5,
  FP_ratio = 2,
  num_iters = 450,
  pair_neighbors = NULL,
  pair_MN = NULL,
  pair_FP = NULL,
  lr = 1,
  apply_pca = TRUE,
  intermediate = FALSE,
  nThreads = parallel::detectCores() - 1
)
}
\arguments{
\item{nThreads}{}
}
\description{
Pairwise Controlled Manifold Approximation (R wrapper)
}
