% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{PaCMAP}
\alias{PaCMAP}
\title{Pairwise Controlled Manifold Approximation (R wrapper)}
\usage{
PaCMAP(object, ...)
}
\arguments{
\item{nThreads}{}
}
\description{
Pairwise Controlled Manifold Approximation (R wrapper)
}
