% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{PaCMAP.Seurat}
\alias{PaCMAP.Seurat}
\title{Pairwise Controlled Manifold Approximation (R wrapper)}
\usage{
\method{PaCMAP}{Seurat}(
  object,
  reduction = "pca",
  reduction_name = "pacmap",
  dims = NULL,
  n_components = 2,
  n_neighbors = 10,
  MN_ratio = 0.5,
  FP_ratio = 2,
  num_iters = 450,
  pair_neighbors = NULL,
  pair_MN = NULL,
  pair_FP = NULL,
  lr = 1,
  apply_pca = TRUE,
  intermediate = FALSE,
  return_seurat = TRUE,
  nThreads = parallel::detectCores() - 1
)
}
\arguments{
\item{return_seurat}{}
}
\description{
Pairwise Controlled Manifold Approximation (R wrapper)
}
