% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{pal_ggplot}
\alias{pal_ggplot}
\title{UMAP Palette using ggplot2 colors}
\usage{
pal_ggplot(object, group_col, jitter = TRUE, comp = 3, alpha = 0.7)
}
\arguments{
\item{comp}{integer setting the color complementarity to be used}
}
\description{
UMAP Palette using ggplot2 colors
}
