% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{pal_hue}
\alias{pal_hue}
\title{UMAP Palette using soft hues}
\usage{
pal_hue(object, group_col, jitter = TRUE, comp = 3, alpha = 0.8)
}
\arguments{
\item{jitter}{integer setting the color complementary to be used}

\item{comp}{integer setting the color complementary to be used}
}
\description{
UMAP Palette using soft hues
}
