% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{pal_rainbow}
\alias{pal_rainbow}
\title{UMAP palette using rainbow colors}
\usage{
pal_rainbow(object, group_col, jitter = TRUE, comp = 3, alpha = 0.7)
}
\arguments{
\item{comp}{integer setting the color complementary to be used}
}
\description{
UMAP palette using rainbow colors
}
