% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{pal_umap}
\alias{pal_umap}
\title{Custom Palette for UMAP}
\usage{
pal_umap(object, group_col, base_col = "#1E90FF", jitter = TRUE)
}
\arguments{
\item{jitter}{randomize the colors}
}
\value{

}
\description{
Custom Palette for UMAP
}
\examples{

}
