% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_helpers.R
\name{pre_processing}
\alias{pre_processing}
\title{Seurat Preprocessing}
\usage{
pre_processing(object, species = "Homo sapiens", nfeatures = 3000, npcs = 50)
}
\arguments{
\item{npcs}{number of principle component dimensions to calculate}
}
\value{

}
\description{
Seurat Preprocessing
}
