% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{rbw_umap}
\alias{rbw_umap}
\title{UMAP palette using rainbow colors}
\usage{
rbw_umap(object, group_col, jitter = TRUE, comp = 3)
}
\arguments{
\item{comp}{integer setting the color complementary to be used}
}
\value{

}
\description{
UMAP palette using rainbow colors
}
