% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read10x.R
\name{read10x}
\alias{read10x}
\title{Read10x v1}
\usage{
read10x(path, return.sce = TRUE)
}
\arguments{
\item{return.sce}{return result as SingleCellExperiment object}
}
\value{

}
\description{
Read10x v1
}
