% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read10x.R
\name{read10x_atlas}
\alias{read10x_atlas}
\title{Read10x Atlas}
\usage{
read10x_atlas(filepaths, project = "scRNAseq", meta = NULL)
}
\arguments{
\item{meta}{meta data to include with the various datasets}
}
\description{
Read10x Atlas
}
