% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_utilities.R
\name{sparse2DT.Seurat}
\alias{sparse2DT.Seurat}
\title{Convert a sparse matrix to a data.table}
\usage{
sparse2DT.Seurat(object)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Convert a sparse matrix to a data.table
}
