% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap-learn.R
\name{umap}
\alias{umap}
\title{UMAP (Uniform Manifold Approximation and Projection for Dimension Reduction)}
\usage{
umap(
  embedding,
  a = NULL,
  angular_rp_forest = FALSE,
  b = NULL,
  force_approximation_algorithm = FALSE,
  init = "spectral",
  learning_rate = 1,
  local_connectivity = 1,
  low_memory = FALSE,
  metric = "euclidean",
  metric_kwds = NULL,
  min_dist = 0.1,
  n_components = 2,
  n_epochs = 200,
  n_neighbors = 15,
  negative_sample_rate = 5,
  output_metric = "euclidean",
  output_metric_kwds = NULL,
  random_state = 42,
  repulsion_strength = 1,
  set_op_mix_ratio = 1,
  spread = 1,
  target_metric = "categorical",
  target_metric_kwds = NULL,
  target_n_neighbors = -1,
  target_weight = 0.5,
  transform_queue_size = 4,
  transform_seed = 42,
  unique = FALSE,
  verbose = TRUE,
  nThreads = parallel::detectCores() - 1
)
}
\arguments{
\item{nThreads}{number of parallel threads to be used}
}
\value{

}
\description{
UMAP (Uniform Manifold Approximation and Projection for Dimension Reduction)
}
