% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_umap.R
\name{umapl}
\alias{umapl}
\title{umap-learn (python implementation of umap)}
\usage{
umapl(object, ...)
}
\arguments{
\item{embedding}{Matrix umap will be performed on.}

\item{nThreads}{number of parallel threads to be used}
}
\description{
umap-learn (python implementation of umap)
}
