% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_umap.R
\name{umapl.Seurat}
\alias{umapl.Seurat}
\title{umap-learn (python implementation of umap)}
\usage{
\method{umapl}{Seurat}(
  object,
  reduction = "pca",
  reduction_name = "umap",
  clustering = FALSE,
  dims = NULL,
  a = 1.662,
  angular_rp_forest = FALSE,
  b = 0.7905,
  densmap = FALSE,
  dens_frac = 0.3,
  dens_var_shift = 0.1,
  output_dens = FALSE,
  disconnection_distance = NULL,
  dens_lambda = 2,
  force_approximation_algorithm = FALSE,
  init = "spectral",
  learning_rate = 1,
  local_connectivity = 1,
  low_memory = FALSE,
  metric = "euclidean",
  metric_kwds = NULL,
  min_dist = 0.1,
  n_components = 2,
  n_epochs = 100,
  n_neighbors = 50,
  negative_sample_rate = 5,
  output_metric = "euclidean",
  output_metric_kwds = NULL,
  random_state = 42,
  repulsion_strength = 1,
  set_op_mix_ratio = 1,
  spread = 1,
  target_metric = "categorical",
  target_metric_kwds = NULL,
  target_n_neighbors = -1,
  target_weight = 0.5,
  transform_queue_size = 4,
  transform_seed = 42,
  unique = FALSE,
  verbose = TRUE,
  nThreads = parallel::detectCores() - 1,
  return_seurat = TRUE
)
}
\arguments{
\item{return_seurat}{}
}
\description{
umap-learn (python implementation of umap)
}
