% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_umap.R
\name{uwot.Seurat}
\alias{uwot.Seurat}
\title{UWOT-UAMP: multithreading capable}
\usage{
\method{uwot}{Seurat}(
  object,
  reduction = "harmony",
  clustering = FALSE,
  dims = NULL,
  n_neighbors = 15,
  n_components = 2,
  metric = "euclidean",
  n_epochs = NULL,
  learning_rate = 1,
  scale = FALSE,
  init = "spectral",
  init_sdev = NULL,
  spread = 1,
  min_dist = 0.01,
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  bandwidth = 1,
  repulsion_strength = 1,
  negative_sample_rate = 5,
  a = NULL,
  b = NULL,
  nn_method = NULL,
  n_trees = 50,
  search_k = 2 * n_neighbors * n_trees,
  approx_pow = FALSE,
  y = NULL,
  target_n_neighbors = n_neighbors,
  target_metric = "euclidean",
  target_weight = 0.5,
  pca = NULL,
  pca_center = TRUE,
  pcg_rand = TRUE,
  fast_sgd = FALSE,
  ret_model = FALSE,
  ret_nn = FALSE,
  ret_extra = c(),
  n_threads = parallel::detectCores() - 1,
  n_sgd_threads = 1,
  grain_size = 1,
  tmpdir = tempdir(),
  verbose = getOption("verbose", TRUE),
  return_seurat = TRUE,
  reduction_name = "umap"
)
}
\arguments{
\item{verbose}{}
}
\description{
UWOT-UAMP: multithreading capable
}
