% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{gg_umap}
\alias{gg_umap}
\title{UMAP Palette using ggplot2 colors}
\usage{
gg_umap(object, group_col, jitter = TRUE, comp = 3)
}
\arguments{
\item{comp}{integer setting the color complementarity to be used}
}
\value{

}
\description{
UMAP Palette using ggplot2 colors
}
