% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan.R
\name{HDBSCAN}
\alias{HDBSCAN}
\title{Hierarchical Density-Based Spatial Clustering of Applications with Noise}
\usage{
HDBSCAN(object, ...)
}
\arguments{
\item{prediction_data}{not sure what this is for. Will update later.}
}
\description{
Hierarchical Density-Based Spatial Clustering of Applications with Noise
}
