% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate_helpers.R
\name{install_python_packages}
\alias{install_python_packages}
\title{Install python packages to use \code{HDBSCAN}, \code{umap-learn}, and \code{PaCMAP}}
\usage{
install_python_packages(method = "auto", conda = "auto")
}
\arguments{
\item{conda}{}
}
\description{
Install python packages to use \code{HDBSCAN}, \code{umap-learn}, and \code{PaCMAP}
}
