% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utilities.R
\name{pre_processing}
\alias{pre_processing}
\title{Seurat Preprocessing}
\usage{
pre_processing(
  object,
  species = "Homo sapiens",
  nfeatures = 3000,
  npcs = 50,
  filters = list(nFeature = c(200, 3000), percent.mt = 20),
  verbose = TRUE
)
}
\arguments{
\item{npcs}{number of principle component dimensions to calculate}
}
\description{
Seurat Preprocessing
}
