% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellphonedb_utilities.R
\name{sparse2DT}
\alias{sparse2DT}
\title{Convert a sparse matrix to a data.table}
\usage{
sparse2DT(matrix)
}
\arguments{
\item{matrix}{sparse matrix to be used}
}
\description{
Convert a sparse matrix to a data.table
}
