% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap-learn.R
\docType{data}
\name{umap}
\alias{umap}
\title{UMAP (Uniform Manifold Approximation and Projection for Dimension Reduction)}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
umap
}
\arguments{
\item{nThreads}{number of parallel threads to be used}
}
\description{
UMAP (Uniform Manifold Approximation and Projection for Dimension Reduction)
}
\keyword{datasets}
